Safari-609.1.11.1/LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/sec-metadata/embed.tentative.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/sec-metadata/resources/record-header.py?file=embed-same-site
Blocked access to external URL https://www.not-web-platform.test:9443/fetch/sec-metadata/resources/record-header.py?file=embed-cross-site
 

FAIL Same-Origin embed assert_not_equals: Empty Sec-Metadata header. got disallowed value ""
FAIL Same-Site embed assert_equals: destination expected (string) "embed" but got (undefined) undefined
FAIL Cross-Site embed assert_equals: destination expected (string) "embed" but got (undefined) undefined
